Revert "search by tag works and renders, search by query works but doesn't re…"
