+++ controllers/index.js
@@ -2,7 +2,7 @@ const router = require('express').Router();
 const apiRoutes = require('./api/');
 const homeRoutes = require('./home-routes.js');
 const dashboardRoutes = require('./dashboard-routes.js');
-const searchRoutes = require('./search-routes.js');
+const searchRoutes = require('./search-routes');
 
 router.use('/api', apiRoutes);
 router.use('/', homeRoutes);
