+++ models/Tag.js
@@ -21,10 +21,6 @@ Tag.init(
             len: [3]
           }
       },
-      tag_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false
-      }
       // user_id: {
       //   type: DataTypes.INTEGER,
       //   references: {
