+++ public/stylesheets/style.css
@@ -1,138 +1,76 @@
 /* RESET? css framework? */
 * {
-  --seafoam: #2fd4be;
-  /* rgb: 47, 212, 190 */
-  --light-seafoam: #adfff4;
-  /* rgb: 173, 255, 244 */
-  --light-grey: #749692;
-  /* rgb: 116, 150, 146 */
-  --dark-seafoam: #071c1a;
-  --med-seafoam: #219485;
+    --seafoam: #2fd4be;
+    /* rgb: 47, 212, 190 */
+    --light-seafoam: #adfff4;
+    /* rgb: 173, 255, 244 */
+    --light-grey: #749692;
+    /* rgb: 116, 150, 146 */
+    --dark-seafoam: #071c1a;
+    --med-seafoam: #219485;
 }
 
 body {
-  font-family: "Kanit", sans-serif;
-  background-color: var(--seafoam);
+    font-family: 'Kanit', sans-serif;
+    background-color: var(--seafoam);
 }
 
 a {
-  text-decoration: none;
+    text-decoration: none;
 }
 
 .main-title {
-  font-family: "Changa One", cursive;
-  font-size: 5vw;
-  color: white;
-  text-align: center;
+    font-family: 'Changa One', cursive;
+    font-size: 5vw;
+    color: white;
+    text-align: center;
 }
 
 .main-title:hover {
-  color: var(--light-grey);
+    color: var(--light-grey);
 }
 
 .nav-item {
-  margin: 0 10px 20px 10px;
-  color: white;
-  font-size: 20px;
+    margin: 0 10px 20px 10px;
+    color: white;
+    font-size: 20px;
 }
 
 .nav-item:hover {
-  color: var(--light-grey);
+    color: var(--light-grey);
 }
 
 .card {
-  padding: 0;
-  margin-top: 8px;
-  margin-bottom: 8px;
-  box-shadow: 5px 5px 2px 1px rgba(116, 150, 146, 0.4);
-  color: var(--dark-seafoam);
+    padding: 0;
+    margin-top: 8px;
+    margin-bottom: 8px;
+    box-shadow: 5px 5px 2px 1px rgba(116, 150, 146, .4);
+    color: var(--dark-seafoam);
 }
 
 .card-title a {
-  color: var(--med-seafoam);
+    color: var(--med-seafoam);
 }
 
 .card-title a:hover {
-  color: var(--light-grey);
+    color: var(--light-grey);
 }
 
 .card-text {
-  font-size: 20px;
+    font-size: 20px;
 }
 
 .card-footer {
-  background-color: rgba(116, 150, 146, 0.2);
-  color: rgba(116, 150, 146, 0.5);
+    background-color: rgba(116, 150, 146, .2);
+    color: rgba(116, 150, 146, .5);
 }
 
 .card-footer .card-text {
-  font-size: 15px;
+    font-size: 15px;
 }
 
 .badge {
-  background-color: var(--light-seafoam);
-  color: var(--dark-seafoam);
-  padding: 5px;
+    background-color: var(--light-seafoam);
+    color: var(--dark-seafoam);
+    padding: 5px;
 }
-
-/* dropdown */
-.dropbtn {
-  background-color: #04aa6d;
-  color: white;
-  padding: 16px;
-  font-size: 16px;
-  border: none;
-  cursor: pointer;
-}
-
-.dropbtn:hover,
-.dropbtn:focus {
-  background-color: #3e8e41;
-}
-
-#myInput {
-  box-sizing: border-box;
-  background-image: url("searchicon.png");
-  background-position: 14px 12px;
-  background-repeat: no-repeat;
-  font-size: 16px;
-  padding: 14px 20px 12px 45px;
-  border: none;
-  border-bottom: 1px solid #ddd;
-}
-
-#myInput:focus {
-  outline: 3px solid #ddd;
-}
-
-.dropdown {
-  position: relative;
-  display: inline-block;
-}
-
-.dropdown-content {
-  display: none;
-  position: absolute;
-  background-color: #f6f6f6;
-  min-width: 230px;
-  overflow: auto;
-  border: 1px solid #ddd;
-  z-index: 1;
-}
-
-.dropdown-content a {
-  color: black;
-  padding: 12px 16px;
-  text-decoration: none;
-  display: block;
-}
-
-.dropdown a:hover {
-  background-color: #ddd;
-}
-
-.show {
-  display: block;
-}
-
-/* end dropdown */
