+++ views/homepage.handlebars
@@ -1,5 +1,3 @@
-
-{{> search tags}}
 <section class="row justify-content-between">
 
     {{#each posts as |post|}}
